# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API文档",
        "title": "历史数据接口文档",
        "contact": {},
        "version": "1.0 版本"
    },
    "basePath": "/app/ipc/api/",
    "paths": {
        "/bridge/count": {
            "post": {
                "description": "查询桥接数量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bridge"
                ],
                "summary": "查询桥接数量",
                "parameters": [
                    {
                        "description": "查询参数",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ParamSearch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-int64"
                        }
                    }
                }
            }
        },
        "/bridge/create": {
            "post": {
                "description": "创建WebRTC-Bridge",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bridge"
                ],
                "summary": "创建WebRTC-Bridge",
                "parameters": [
                    {
                        "description": "桥接信息",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Bridge"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Bridge"
                        }
                    }
                }
            }
        },
        "/bridge/export": {
            "get": {
                "description": "导出WebRTC-Bridge",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "product"
                ],
                "summary": "导出WebRTC-Bridge",
                "responses": {}
            }
        },
        "/bridge/import": {
            "post": {
                "description": "导入WebRTC-Bridge",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "导入WebRTC-Bridge",
                "parameters": [
                    {
                        "type": "file",
                        "description": "压缩包",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-int64"
                        }
                    }
                }
            }
        },
        "/bridge/list": {
            "get": {
                "description": "查询WebRTC-Bridge",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bridge"
                ],
                "summary": "查询WebRTC-Bridge",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "skip",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyList-types_Bridge"
                        }
                    }
                }
            }
        },
        "/bridge/search": {
            "post": {
                "description": "查询WebRTC-Bridge",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bridge"
                ],
                "summary": "查询WebRTC-Bridge",
                "parameters": [
                    {
                        "description": "查询参数",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ParamSearch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyList-types_Bridge"
                        }
                    }
                }
            }
        },
        "/bridge/{id}": {
            "get": {
                "description": "获取WebRTC-Bridge",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bridge"
                ],
                "summary": "获取WebRTC-Bridge",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "桥接ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Bridge"
                        }
                    }
                }
            },
            "post": {
                "description": "修改WebRTC-Bridge",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bridge"
                ],
                "summary": "修改WebRTC-Bridge",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "桥接ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "桥接信息",
                        "name": "bridge",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Bridge"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Bridge"
                        }
                    }
                }
            }
        },
        "/bridge/{id}/delete": {
            "get": {
                "description": "删除WebRTC-Bridge",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bridge"
                ],
                "summary": "删除WebRTC-Bridge",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "桥接ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Bridge"
                        }
                    }
                }
            }
        },
        "/bridge/{id}/disable": {
            "get": {
                "description": "禁用WebRTC-Bridge",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bridge"
                ],
                "summary": "禁用WebRTC-Bridge",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "桥接ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Bridge"
                        }
                    }
                }
            }
        },
        "/bridge/{id}/enable": {
            "get": {
                "description": "启用WebRTC-Bridge",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bridge"
                ],
                "summary": "启用WebRTC-Bridge",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "桥接ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Bridge"
                        }
                    }
                }
            }
        },
        "/camera/create": {
            "post": {
                "description": "创建摄像头",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "camera"
                ],
                "summary": "创建摄像头",
                "parameters": [
                    {
                        "description": "摄像头信息",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Camera"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Camera"
                        }
                    }
                }
            }
        },
        "/camera/export": {
            "get": {
                "description": "导出摄像头",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "product"
                ],
                "summary": "导出摄像头",
                "responses": {}
            }
        },
        "/camera/import": {
            "post": {
                "description": "导入摄像头",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "导入摄像头",
                "parameters": [
                    {
                        "type": "file",
                        "description": "压缩包",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-int64"
                        }
                    }
                }
            }
        },
        "/camera/list": {
            "get": {
                "description": "查询摄像头",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "camera"
                ],
                "summary": "查询摄像头",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "skip",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyList-types_Camera"
                        }
                    }
                }
            }
        },
        "/camera/search": {
            "post": {
                "description": "查询摄像头",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "camera"
                ],
                "summary": "查询摄像头",
                "parameters": [
                    {
                        "description": "查询参数",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ParamSearch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyList-types_Camera"
                        }
                    }
                }
            }
        },
        "/camera/{id}": {
            "get": {
                "description": "获取摄像头",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "camera"
                ],
                "summary": "获取摄像头",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "摄像头ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Camera"
                        }
                    }
                }
            },
            "post": {
                "description": "修改摄像头",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "camera"
                ],
                "summary": "修改摄像头",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "摄像头ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "摄像头信息",
                        "name": "camera",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Camera"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Camera"
                        }
                    }
                }
            }
        },
        "/camera/{id}/delete": {
            "get": {
                "description": "删除摄像头",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "camera"
                ],
                "summary": "删除摄像头",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "摄像头ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Camera"
                        }
                    }
                }
            }
        },
        "/camera/{id}/disable": {
            "get": {
                "description": "禁用摄像头",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "camera"
                ],
                "summary": "禁用摄像头",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "摄像头ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Camera"
                        }
                    }
                }
            }
        },
        "/camera/{id}/enable": {
            "get": {
                "description": "启用摄像头",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "camera"
                ],
                "summary": "启用摄像头",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "摄像头ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Camera"
                        }
                    }
                }
            }
        },
        "/job/count": {
            "post": {
                "description": "查询摄像头数量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "camera"
                ],
                "summary": "查询摄像头数量",
                "parameters": [
                    {
                        "description": "查询参数",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ParamSearch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-int64"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ParamSearch": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "object",
                    "additionalProperties": true
                },
                "keyword": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "skip": {
                    "type": "integer"
                },
                "sort": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                }
            }
        },
        "api.ReplyData-int64": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "api.ReplyData-types_Bridge": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/types.Bridge"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "api.ReplyData-types_Camera": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/types.Camera"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "api.ReplyList-types_Bridge": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Bridge"
                    }
                },
                "error": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "api.ReplyList-types_Camera": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Camera"
                    }
                },
                "error": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "types.Bridge": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "types.Camera": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ipc_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}